+++ docs/Changelog.md
@@ -6,6 +6,7 @@
 * enh: dict in a dataclass support
 * fix: config file with str-attribute missing clash
 * fix: SelectTag multiple choice without default value resilience
+* fix: dynamic Literal in Annotated
 
 ## 1.1.1 (2025-09-18)
 * enh: `list[tuple]` support (along with `list[tuple[int]]` and `list[tuple[int, ...]]`)